добавляет ДЗ модуля 7 задание 1
